Fix bugs
Fixed some bugs occuring with 'optimizationXZ' for some test cases.
The bug was due to the decomposition introducing terms that would
result in a reflection, but they were not summed correctly. A summation
of gates was added to '_swap_Y_for_Z' to remove redundancies like
Y(theta)Y(gamma).