Add 'hardware_running_time'
Added a 'hardware_running_time' to return the running time of the
circuit in the hardware. It receives to parameters as inputs 'lengthZ'
and 'lengthX' which are the expected running time of the Z gate and
X gate respectively. Then, the method loops through each gate and sum
the time each one of them take.